backup/webapp/requirements.txt
# api
cos-python-sdk-v5
streamlit
pillow==9.1.0